requirements.txt
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
filelock==3.15.4
fsspec==2024.6.0
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.4.3
soundfile==0.12.1
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
urllib3==2.2.2
